pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ka-chat-bot"
version = "0.1.0"
description = "Databricks Chat Bot Application"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "pydantic>=2.5.0",
    "python-dotenv==1.0.0",
    "databricks-sdk==0.62.0",
    "aiofiles==23.2.1",
    "backoff==2.2.1",
    "gunicorn==23.0.0",
    "httpx==0.25.2",
    "sqlalchemy==2.0.42",
    "asyncpg>=0.29.0",
    "greenlet>=2.0.0",
    "langgraph-checkpoint-postgres==2.0.23",
    "psycopg==3.2.9",
    "psycopg-binary==3.2.9",
    "psycopg2-binary==2.9.10",
    "psycopg-pool==3.2.6",
    "unitycatalog-langchain==0.2.0",
    "langchain>=0.3.27",
    "langchain-community",
    "langgraph==0.6.4",
    "langchain-openai",
    "databricks-langchain",
    "unitycatalog-ai[databricks]",
    "mlflow[databricks]>=3.2.0",
    "uuid>=1.30",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

[tool.ruff]
target-version = "py310"
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "uuid>=1.30",
]
